requirements.txt
azure-functions
grafana_backup==1.1.3